In Summary Tab Add the use Query
